receive Data BookTask

+++ Contract/models/PublisherModel.cs
@@ -21,6 +21,6 @@ namespace Contract.models
         public DateTime DateOfBirth { get; set; }
         [Required]
         [Range(2000, 5000, ErrorMessage = "it must be between 2000 and 5000 $")]
-        public int Salery { get; set; }
+        public float Salery { get; set; }
     }
 }
